feat: update legacy test configs
