--- tests-integrations/config-recommended/.eslintrc.js
@@ -1,6 +1,0 @@
-'use strict'
-
-module.exports = {
-  root: true,
-  extends: [ 'plugin:@intlify/svelte/recommended'],
-}
