+++ tests-integrations/legacy-config/.eslintrc.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  root: true,
+  extends: ['plugin:@intlify/svelte/recommended']
+}
