+++ tests-integrations/legacy-config/.npmrc
