+++ tests-integrations/legacy-config/.vscode/settings.json
