+++ tests-integrations/legacy-config/src/a.svelte
